tests/submission/Microsoft/semantic/figure.html

<!doctype html>
<html>
<head>
 <meta content="text/html; charset=utf-8" http-equiv="content-type" />
 <title>verify FIGURE is a HTMLElement</title>
  <script src="/resources/testharness.js"></script>
  <script src="http://www.w3c-test.org/resources/testharnessreport.js"></script>
</head>
<body>
 <div id="log"></div>
 <figure id="figure">
  <figcaption>This is just an example of a FIGCAPTION in a FIGURE:</figcaption>
 </figure>

 <script type="text/javascript">
  test(function()
  {
     assert_array_equals( "[object HTMLElement]",
                          document.getElementById("figure").toString() )
  })
 </script>
</body>
</html>
